+++ src/mcp/tools/firestore/delete_document.ts
@@ -10,11 +10,10 @@ export const delete_document = tool(
     description:
       "Deletes a Firestore documents from a database in the current project by full document paths. Use this if you know the exact path of a document.",
     inputSchema: z.object({
-      // TODO: Support configurable database
-      // database: z
-      //   .string()
-      //   .optional()
-      //   .describe("Database id to use. Defaults to `(default)` if unspecified."),
+      database: z
+        .string()
+        .optional()
+        .describe("Database id to use. Defaults to `(default)` if unspecified."),
       path: z
         .string()
         .describe(
@@ -30,14 +29,15 @@ export const delete_document = tool(
       requiresProject: true,
     },
   },
-  async ({ path }, { projectId }) => {
-    // database ??= "(default)";
-    const { documents, missing } = await getDocuments(projectId, [path]);
+  async ({ path, database }, { projectId }) => {
+    const { documents, missing } = await getDocuments(projectId, [path], database);
     if (missing.length > 0 && documents && documents.length === 0) {
       return mcpError(`None of the specified documents were found in project '${projectId}'`);
     }
 
-    const firestoreDelete = new FirestoreDelete(projectId, path, { databaseId: "(default)" });
+    const firestoreDelete = new FirestoreDelete(projectId, path, {
+      databaseId: database ?? "(default)",
+    });
 
     await firestoreDelete.execute();