+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Improved GetDatabase API call caching for Firestore function deployments. (#8681)
 - Increased timeout for linking CloudSQL instances to Data Connect.
+- Correctly support '--database' flag in `firestore:delete`. (#6753)